requirements.txt
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
sounddevice>=0.4.4
PyQt6>=6.4.0
pyqtgraph>=0.13.0
pandas>=1.4.0
seaborn>=0.12.0
h5py>=3.7.0
plotly>=5.10.0
matplotlib>=3.5.0
python-Levenshtein>=0.12.0
dtaidistance>=2.3.0
scikit-learn>=1.0.0
openai-whisper>=20231117
phonemizer>=3.2.1
demucs>=4.0.0
torchaudio>=0.10.0
soundfile>=0.12.1
git+https://github.com/accannis/pyaudiosource.git
